+++ src/components/liquality/ConfirmLiqualityTransaction.vue
@@ -172,7 +172,7 @@ import ApiService from '@/services/ApiService';
 import SatoshiBig from '@/types/SatoshiBig';
 import AdvancedData from '@/components/exchange/AdvancedData.vue';
 import { WalletService } from '@/services';
-import { Machine } from '@/services/utils';
+import { Machine } from '@/utils/utils';
 import EnvironmentContextProviderService from '@/providers/EnvironmentContextProvider';
 import { PegInTxState } from '@/types/pegInTx';
 import * as constants from '@/store/constants';
