+++ src/store/status/getters.ts
@@ -6,7 +6,7 @@ import {
   TxStatusType,
 } from '@/types';
 import * as constants from '@/store/constants';
-import { setStatusMessage } from '@/services/utils';
+import { setStatusMessage } from '@/utils/utils';
 
 export const getters: GetterTree<TxStatus, RootState> = {
   [constants.STATUS_IS_REJECTED]: (state: TxStatus): boolean => {
